Move authorization to pundit
Additional changes:
* Fixed "banned for permanent" in records
* Added columns on user profiles for stats when user has no uploads
* Made name change request reasons visible to moderators
* Changed `GET /users/:id/edit` to `GET /users/edit` & `PATCH /users/:id` to `POST /users/update`

I really should have done this process in steps.

+++ app/controllers/forum_topics_controller.rb
@@ -2,32 +2,15 @@
 
 class ForumTopicsController < ApplicationController
   respond_to :html, :json
-  before_action :member_only, except: %i[index show]
-  before_action :moderator_only, only: [:unhide]
-  before_action :admin_only, only: [:destroy]
-  before_action :normalize_search, only: :index
   before_action :load_topic, only: %i[edit show update destroy hide unhide subscribe unsubscribe]
-  before_action :check_min_level, only: %i[show edit update destroy hide unhide subscribe unsubscribe]
   skip_before_action :api_check
-
-  def new
-    @forum_topic = ForumTopic.new(forum_topic_params)
-    @forum_topic.original_post = ForumPost.new(forum_topic_params[:original_post_attributes])
-    respond_with(@forum_topic)
-  end
-
-  def edit
-    check_privilege(@forum_topic)
-    respond_with(@forum_topic)
-  end
 
   def index
     params[:search] ||= {}
-    params[:search][:order] ||= "sticky" if request.format == Mime::Type.lookup("text/html")
+    params[:search][:order] ||= "sticky" if request.format.html?
 
-    @query = ForumTopic.permitted.active.search(search_params)
-    @query = ForumTopic.permitted.search(search_params) if CurrentUser.is_moderator?
-    @forum_topics = @query.paginate(params[:page], limit: per_page, search_count: params[:search])
+    @query = authorize(ForumTopic).visible(CurrentUser.user).search(search_params(ForumTopic))
+    @forum_topics = @query.paginate(params[:page], limit: params[:limit] || 50, search_count: params[:search])
 
     respond_with(@forum_topics) do |format|
       format.html do
@@ -39,8 +22,20 @@ class ForumTopicsController < ApplicationController
     end
   end
 
+  def new
+    @forum_topic = authorize(ForumTopic.new(permitted_attributes(ForumTopic)))
+    @forum_topic.original_post = ForumPost.new(permitted_attributes(ForumTopic)[:original_post_attributes])
+    respond_with(@forum_topic)
+  end
+
+  def edit
+    authorize(@forum_topic)
+    respond_with(@forum_topic)
+  end
+
   def show
-    if request.format == Mime::Type.lookup("text/html")
+    authorize(@forum_topic)
+    if request.format.html?
       @forum_topic.mark_as_read!(CurrentUser.user)
     end
     @forum_posts = ForumPost.includes(topic: [:category]).search(topic_id: @forum_topic.id).reorder("forum_posts.id").paginate(params[:page])
@@ -49,25 +44,27 @@ class ForumTopicsController < ApplicationController
   end
 
   def create
-    @forum_topic = ForumTopic.create(forum_topic_params)
+    @forum_topic = authorize(ForumTopic.new(permitted_attributes(ForumTopic)))
+    @forum_topic.save
     respond_with(@forum_topic)
   end
 
   def update
-    check_privilege(@forum_topic)
-    @forum_topic.assign_attributes(forum_topic_params)
+    authorize(@forum_topic)
+    @forum_topic.assign_attributes(permitted_attributes(ForumTopic))
     @forum_topic.save(touch: false)
     respond_with(@forum_topic)
   end
 
   def destroy
-    check_privilege(@forum_topic)
+    authorize(@forum_topic)
     @forum_topic.destroy
-    flash[:notice] = "Topic deleted"
+    notice("Topic deleted")
+    respond_with(@forum_topic, location: forum_topics_path)
   end
 
   def hide
-    check_privilege(@forum_topic)
+    authorize(@forum_topic)
     @forum_topic.hide!
     if @forum_topic.errors.any?
       respond_with(@forum_topic) do |format|
@@ -76,19 +73,20 @@ class ForumTopicsController < ApplicationController
         end
       end
     else
-      flash[:notice] = "Topic hidden"
+      notice("Topic hidden")
       respond_with(@forum_topic)
     end
   end
 
   def unhide
-    check_privilege(@forum_topic)
+    authorize(@forum_topic)
     @forum_topic.unhide!
     flash[:notice] = "Topic unhidden"
     respond_with(@forum_topic)
   end
 
   def mark_all_as_read
+    authorize(ForumTopic)
     CurrentUser.user.update_attribute(:last_forum_read_at, Time.now)
     ForumTopicVisit.prune!(CurrentUser.user)
     respond_to do |format|
@@ -98,6 +96,7 @@ class ForumTopicsController < ApplicationController
   end
 
   def subscribe
+    authorize(@forum_topic)
     subscription = ForumSubscription.where(forum_topic_id: @forum_topic.id, user_id: CurrentUser.user.id).first
     unless subscription
       ForumSubscription.create(forum_topic_id: @forum_topic.id, user_id: CurrentUser.user.id, last_read_at: @forum_topic.updated_at)
@@ -106,6 +105,7 @@ class ForumTopicsController < ApplicationController
   end
 
   def unsubscribe
+    authorize(@forum_topic)
     subscription = ForumSubscription.where(forum_topic_id: @forum_topic.id, user_id: CurrentUser.user.id).first
     if subscription
       subscription.destroy
@@ -115,40 +115,7 @@ class ForumTopicsController < ApplicationController
 
   private
 
-  def per_page
-    params[:limit] || 40
-  end
-
-  def normalize_search
-    if params[:title_matches]
-      params[:search] ||= {}
-      params[:search][:title_matches] = params.delete(:title_matches)
-    end
-
-    if params[:title]
-      params[:search] ||= {}
-      params[:search][:title] = params.delete(:title)
-    end
-  end
-
-  def check_privilege(forum_topic)
-    if !forum_topic.editable_by?(CurrentUser.user)
-      raise(User::PrivilegeError)
-    end
-  end
-
   def load_topic
     @forum_topic = ForumTopic.includes(:category).find(params[:id])
-  end
-
-  def check_min_level
-    raise(User::PrivilegeError.new) unless @forum_topic.visible?(CurrentUser.user)
-  end
-
-  def forum_topic_params
-    permitted_params = [:title, :category_id, { original_post_attributes: %i[id body] }]
-    permitted_params += %i[is_sticky is_locked] if CurrentUser.is_moderator?
-
-    params.fetch(:forum_topic, {}).permit(permitted_params)
   end
 end
